Adjusting page type

+++ src/Contracts/Services/Warehouse/Query.cs
@@ -4,7 +4,7 @@ namespace Contracts.Services.Warehouse;
 
 public static class Query
 {
-    public record GetInventories(int Limit, int Offset) : Message, IQuery;
+    public record GetInventories(ushort Limit, ushort Offset) : Message, IQuery;
 
-    public record GetInventoryItems(Guid InventoryId, int Limit, int Offset) : Message(CorrelationId: InventoryId), IQuery;
+    public record GetInventoryItems(Guid InventoryId, ushort Limit, ushort Offset) : Message(CorrelationId: InventoryId), IQuery;
 }
